+++ src/kairos/workspace/domain/workspace.type.ts
