+++ src/kairos/workspace/interface/http/create-workspace-response.ts
@@ -1,7 +1,7 @@
 import { type ResultHttpMapping } from '@/interface/http/result-to-http';
 import { HTTP_CONFLICT, HTTP_CREATED } from '@/interface/http/status-code';
 import type { WorkspaceSlugConflictError } from '@/kairos/workspace/domain/errors';
-import type { Workspace } from '@/kairos/workspace/domain/workspace';
+import type { Workspace } from '@/kairos/workspace/domain/workspace.type';
 
 export const createWorkspaceResponse: ResultHttpMapping<Workspace, WorkspaceSlugConflictError> = {
   success: {
